requirements.txt
flask
jinja2
pandas
requests
requests-oauthlib
cryptography
pyjwt
redis
